Second Phase and Three main tests done.

+++ src/main/java/org/example/investmentfunds/investment/style/InvestmentStyle.java
@@ -2,9 +2,10 @@ package org.example.investmentfunds.investment.style;
 
 import org.example.investmentfunds.model.FundType;
 
+import java.math.BigDecimal;
 import java.util.Map;
 
 public interface InvestmentStyle {
 
-    Map<FundType, Double> getInvestmentRatioMap();
+    Map<FundType, BigDecimal> getInvestmentRatioMap();
 }
